[env](compile) add compiler check and fix some clang-tidy suggestions for vec/exec/format (#45881)
### What problem does this PR solve?
1. Enable compiler check in vec/exec/format files
2. Fix clang-tidy by `clang-tidy --fix--errors`

+++ be/src/vec/data_types/serde/data_type_jsonb_serde.h
@@ -56,7 +56,7 @@ public:
                                           const FormatOptions& options) const override;
 
     Status deserialize_column_from_json_vector(IColumn& column, std::vector<Slice>& slices,
-                                               int* num_deserialized,
+                                               uint64_t* num_deserialized,
                                                const FormatOptions& options) const override;
 
     Status write_column_to_orc(const std::string& timezone, const IColumn& column,
